Moving utils and serviceUtils into payment module.
Renaming serviceUtils to handleRequestError since that’s the only thing in it.
